+++ Sources/ZcashLightClientKit/ClosureSynchronizer.swift
@@ -79,12 +79,8 @@ public protocol ClosureSynchronizer {
 
     func getTransparentBalance(accountIndex: Int, completion: @escaping (Result<WalletBalance, Error>) -> Void)
 
-    @available(*, deprecated, message: "This function will be removed soon, use the one returning a `Zatoshi` value instead")
-    func getShieldedBalance(accountIndex: Int) -> Int64
     func getShieldedBalance(accountIndex: Int) -> Zatoshi
 
-    @available(*, deprecated, message: "This function will be removed soon, use the one returning a `Zatoshi` value instead")
-    func getShieldedVerifiedBalance(accountIndex: Int) -> Int64
     func getShieldedVerifiedBalance(accountIndex: Int) -> Zatoshi
 
     /*