+++ Sources/ZcashLightClientKit/Rust/ZcashRustBackendWelding.swift
@@ -156,17 +156,6 @@ protocol ZcashRustBackendWelding {
     /// - parameter dbData: location of the data db file
     /// - parameter idNote: note_id of note where the memo is located
     /// - parameter networkType: network type of this key
-    @available(*, deprecated, message: "This function will be deprecated soon. Use `getReceivedMemo(dbData:idNote:networkType)` instead")
-    static func getReceivedMemoAsUTF8(
-        dbData: URL,
-        idNote: Int64,
-        networkType: NetworkType
-    ) -> String?
-
-    /// get received memo from note
-    /// - parameter dbData: location of the data db file
-    /// - parameter idNote: note_id of note where the memo is located
-    /// - parameter networkType: network type of this key
     static func getReceivedMemo(
         dbData: URL,
         idNote: Int64,
@@ -178,17 +167,6 @@ protocol ZcashRustBackendWelding {
     /// - Returns a `SaplingAddress` if any
     /// - Throws `receiverNotFound` when the receiver is not found. `invalidUnifiedAddress` if the UA provided is not valid
     static func getSaplingReceiver(for uAddr: UnifiedAddress) throws -> SaplingAddress?
-
-    /// get sent memo from note
-    /// - parameter dbData: location of the data db file
-    /// - parameter idNote: note_id of note where the memo is located
-    /// - parameter networkType: network type of this key
-    @available(*, deprecated, message: "This function will be deprecated soon. Use `getSentMemo(dbData:idNote:networkType)` instead")
-    static func getSentMemoAsUTF8(
-        dbData: URL,
-        idNote: Int64,
-        networkType: NetworkType
-    ) -> String?
 
     /// get sent memo from note
     /// - parameter dbData: location of the data db file