+++ Sources/ZcashLightClientKit/Synchronizer.swift
@@ -259,15 +259,7 @@ public protocol Synchronizer: AnyObject {
     func getTransparentBalance(accountIndex: Int) async throws -> WalletBalance
 
     /// Returns the shielded total balance (includes verified and unverified balance)
-    @available(*, deprecated, message: "This function will be removed soon, use the one returning a `Zatoshi` value instead")
-    func getShieldedBalance(accountIndex: Int) -> Int64
-
-    /// Returns the shielded total balance (includes verified and unverified balance)
     func getShieldedBalance(accountIndex: Int) -> Zatoshi
-
-    /// Returns the shielded verified balance (anchor is 10 blocks back)
-    @available(*, deprecated, message: "This function will be removed soon, use the one returning a `Zatoshi` value instead")
-    func getShieldedVerifiedBalance(accountIndex: Int) -> Int64
 
     /// Returns the shielded verified balance (anchor is 10 blocks back)
     func getShieldedVerifiedBalance(accountIndex: Int) -> Zatoshi